backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
asyncpg==0.29.0
SQLAlchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
packaging==24.0
